# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Product Hackathon API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/entities/{entity_id}/graph": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Entity Graph",
        "operationId": "entity_graph_api_v1_entities__entity_id__graph_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2,
              "minimum": 1,
              "default": 1,
              "title": "Depth"
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/EntityType"
              },
              "title": "Types"
            }
          },
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period From"
            }
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 40,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_GraphView_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Search Catalog",
        "operationId": "search_catalog_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/EntityType"
              },
              "title": "Types"
            }
          },
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period From"
            }
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Page_SearchItemView__"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/timeline/events": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Timeline Events",
        "operationId": "timeline_events_api_v1_timeline_events_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period From"
            }
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Page_TimelineEventView__"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
...
      "ApiResponse_GraphView_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphView"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ApiResponse[GraphView]"
      },
...
      "ApiResponse_NoneType_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "data": {
            "type": "null",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ApiResponse[NoneType]"
      },
...
      "ApiResponse_Page_SearchItemView__": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_SearchItemView_"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ApiResponse[Page[SearchItemView]]"
      },
...
      "ApiResponse_Page_TimelineEventView__": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_TimelineEventView_"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ApiResponse[Page[TimelineEventView]]"
      },
...
      "Coordinates": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "Coordinates"
      },
...
      "EntityType": {
        "type": "string",
        "enum": [
          "settlement",
          "person",
          "event",
          "landmark",
          "natural_object",
          "cultural_object",
          "organization",
          "university_object",
          "artifact"
        ],
        "title": "EntityType"
      },
...
      "GraphCenterView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "title": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "title"
        ],
        "title": "GraphCenterView"
      },
      "GraphEdgeView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "type": {
            "$ref": "#/components/schemas/RelationType"
          },
          "title": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "description": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "sources_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sources Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_id",
          "target_id",
          "type",
          "title",
          "description",
          "sources_count"
        ],
        "title": "GraphEdgeView"
      },
      "GraphNodeView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "title": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "relations_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Relations Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "relations_count"
        ],
        "title": "GraphNodeView"
      },
      "GraphView": {
        "properties": {
          "center": {
            "$ref": "#/components/schemas/GraphCenterView"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeView"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeView"
            },
            "type": "array",
            "title": "Edges"
          },
          "hidden_nodes_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Hidden Nodes Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "center",
          "nodes",
          "edges",
          "hidden_nodes_count"
        ],
        "title": "GraphView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageMeta": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total"
        ],
        "title": "PageMeta"
      },
...
      "Page_SearchItemView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchItemView"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[SearchItemView]"
      },
...
      "Page_TimelineEventView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineEventView"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[TimelineEventView]"
      },
...
      "RelationType": {
        "type": "string",
        "enum": [
          "born_in",
          "lived_in",
          "worked_in",
          "studied_in",
          "taught_at",
          "participated_in",
          "located_in",
          "part_of",
          "created_by",
          "described_in",
          "connected_with",
          "connected_with_chgu"
        ],
        "title": "RelationType"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "ResponseMeta"
      },
...
      "SearchItemView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "title": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "subtitle": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "relations_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Relations Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "subtitle",
          "cover_url",
          "coordinates",
          "relations_count"
        ],
        "title": "SearchItemView"
      },
...
      "TimelineEventView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "short_description": {
            "$ref": "#/components/schemas/modules__catalog__schemas__LocalizedText"
          },
          "period_from": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period From"
          },
          "period_to": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period To"
          },
          "coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "short_description",
          "period_from",
          "period_to",
          "coordinates"
        ],
        "title": "TimelineEventView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "modules__catalog__schemas__LocalizedText": {
        "properties": {
          "ru": {
            "type": "string",
            "minLength": 1,
            "title": "Ru"
          },
          "ce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ce"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ru",
          "ce"
        ],
        "title": "LocalizedText"
      },
...
    }
  }
}